% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{air_temp_daily_dd}
\alias{air_temp_daily_dd}
\title{Daily average air temperature and degree days for the Delta region}
\format{
data frame with 2,556 rows and 4 columns
\describe{
\item{Year}{Calendar year of the measurement}
\item{Date}{Calendar date of the measurement in yyyy-mm-dd}
\item{AirTemp_Avg}{Daily average air temperature in degrees Celsius of
three continuous weather stations located near the Delta}
\item{AirTemp_DD}{Daily accumulated degree days above 19C for the averaged
air temperature data at three stations located near the Delta for each
\code{Year}. Calculated with the \code{\link[pollen:gdd]{pollen::gdd()}} function.}
}
}
\source{
California Department of Water Resources (DWR) California Data
Exchange Center (CDEC): \url{https://cdec.water.ca.gov/}
}
\usage{
air_temp_daily_dd
}
\description{
A data set containing daily average air temperatures and degree
days above 19 degrees Celsius for the region near the upper San Francisco
Estuary (Delta) from 2015-2021. Values are based on averages of three
continuous weather observation stations located near the Delta: San Joaquin
River McCune station (SJR), Harvey O' Banks Pumping Plant (HBP), and San
Joaquin River at Mossdale (MSD). Used in the analyses for the Spring-Summer
version of the 2022 HABs/Weeds report for the Emergency Drought Barrier.

All air temperature data was collected by DWR and was downloaded from the
California Data Exchange Center (CDEC).
}
\keyword{datasets}
