% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cont_chla_daily}
\alias{cont_chla_daily}
\title{Daily values of continuous chlorophyll fluorescence data from
2020-2021}
\format{
data frame with 7,611 rows and 6 columns
\describe{
\item{Station}{Location where measurement was collected}
\item{Region}{Region designation of \code{Station} for the Emergency Drought
Barrier analysis. Either Central Delta, Sacramento, or San Joaquin.}
\item{Year}{Calendar year of the value}
\item{Date}{Calendar date of the value}
\item{Chla_avg}{Daily average chlorophyll fluorescence value in ug/L}
\item{Chla_med}{Daily median chlorophyll fluorescence value in ug/L}
}
}
\source{
\itemize{
\item California Department of Water Resources (DWR) Water Data Library (WDL):
\url{https://wdl.water.ca.gov/WaterDataLibrary/}
\item USGS National Water Information System (NWIS):
\url{https://nwis.waterdata.usgs.gov/nwis}
}
}
\usage{
cont_chla_daily
}
\description{
A data set containing daily average and median values of
continuous chlorophyll fluorescence data collected by DWR and USGS from
2020 - 2021 for the stations within the designated Emergency Drought
Barrier (EDB) regions. Used in the analyses for the EDB.

All chlorophyll data collected by DWR-NCRO (DWR's North Central Regional
Office) was downloaded from the Water Data Library (WDL). Data collected by
DWR-CEMP (DWR's Continuous Environmental Monitoring Program) was acquired
through direct data requests. And data from USGS were downloaded directly
from the National Water Information System (NWIS) using the \code{dataRetrieval} R
package. Some of the data from USGS is considered provisional.
}
\keyword{datasets}
