% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cont_wq_daily}
\alias{cont_wq_daily}
\title{Daily values of continuous water quality data for stations near Franks
Tract}
\format{
data frame with 17,552 rows and 11 columns
\describe{
\item{Source}{Name of the source dataset. Is one of the following: DWR_CEMP
(DWR's Continuous Environmental Monitoring Program), DWR_NCRO (DWR's North
Central Region Office), or USGS}
\item{Station}{Location where measurement was collected}
\item{Date}{Calendar date of the measurement in yyyy-mm-dd}
\item{WaterTemp_Mean}{Daily average water temperature in degrees Celcius}
\item{WaterTemp_Max}{Daily maximum water temperature in degrees Celcius}
\item{DO_Mean}{Daily average dissolved oxygen in milligrams per liter}
\item{DO_Max}{Daily maximum dissolved oxygen in milligrams per liter}
\item{pH_Mean}{Daily average pH in pH units}
\item{pH_Max}{Daily maximum pH in pH units}
\item{Chla_Mean}{Daily average chlorophyll fluorescence in micrograms per
liter}
\item{Chla_Max}{Daily maximum chlorophyll fluorescence in micrograms per
liter}
}
}
\source{
\itemize{
\item California Department of Water Resources (DWR) Water Data Library (WDL):
\url{https://wdl.water.ca.gov/WaterDataLibrary/}
\item USGS National Water Information System (NWIS):
\url{https://nwis.waterdata.usgs.gov/nwis}
}
}
\usage{
cont_wq_daily
}
\description{
A data set containing daily mean and maximum values of various
continuous water quality parameters measured in the water in-situ at the
stations near Franks Tract located in the upper San Francisco Estuary
(Delta). Water quality parameters include Water Temperature, Dissolved
Oxygen, pH, and Chlorophyll Fluorescence. Data was collected by DWR and
USGS from 2015-2021. Used in the analyses for the Spring-Summer version of
the 2022 HABs/Weeds report for the Emergency Drought Barrier.

All water quality data collected by DWR_NCRO (DWR's North Central Region
Office) was downloaded from the Water Data Library (WDL). Data collected by
DWR_CEMP (DWR's Continuous Environmental Monitoring Program) was acquired
through direct data requests. And data from USGS were downloaded from the
National Water Information System (NWIS).
}
\keyword{datasets}
