% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cont_wq_daily_max_frk}
\alias{cont_wq_daily_max_frk}
\title{Daily maximums of continuous water quality data for the Franks Tract
station (FRK)}
\format{
data frame with 2,376 rows and 4 columns
\describe{
\item{Date}{Calendar date of the measurement in yyyy-mm-dd}
\item{DO_PerSat}{Daily maximum dissolved oxygen in percent saturation.
Values were converted from the recorded dissolved oxygen concentration in
mg/L to percent saturation using the USGS method implemented from the
\href{https://cran.r-project.org/web/packages/rMR/index.html}{\code{rMR}} package.}
\item{pH}{Daily maximum pH in pH units}
\item{Chla}{Daily maximum chlorophyll fluorescence in micrograms per liter}
}
}
\source{
Acquired through direct data requests
}
\usage{
cont_wq_daily_max_frk
}
\description{
A data set containing daily maximum values of various continuous
water quality parameters measured in the water in-situ at the the Franks
Tract station (FRK) located in the upper San Francisco Estuary (Delta).
Water quality parameters include Dissolved Oxygen (as percent saturation),
pH, and Chlorophyll Fluorescence. Data was collected from 2015-2021. Used
in the analyses for the Spring-Summer version of the 2022 HABs/Weeds report
for the Emergency Drought Barrier.

Water quality data was collected by DWR_CEMP (DWR's Continuous
Environmental Monitoring Program) and was acquired through direct data
requests.
}
\keyword{datasets}
