% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDBdata-package.R
\docType{package}
\name{EDBdata-package}
\alias{EDBdata}
\alias{EDBdata-package}
\title{EDBdata: Emergency Drought Barrier Data}
\description{
Contains data sets used in the Emergency Drought Barrier (EDB) analysis. This data package also contains useful metadata describing the data sets and the processing scripts used to derive the data sets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mountaindboz/EDBdata}
  \item Report bugs at \url{https://github.com/mountaindboz/EDBdata/issues}
}

}
\author{
\strong{Maintainer}: Dave Bosworth \email{David.Bosworth@water.ca.gov} (\href{https://orcid.org/0000-0003-0740-3390}{ORCID})

}
\keyword{internal}
