% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hab_incidents}
\alias{hab_incidents}
\title{CyanoHAB incident reports in the upper San Francisco Estuary}
\format{
data frame with 50 rows and 4 columns
\describe{
\item{Date}{Calendar date of the incident in yyyy-mm-dd}
\item{Latitude}{Latitude of incident location in Decimal Degrees (WGS 84
Datum)}
\item{Longitude}{Longitude of incident location in Decimal Degrees (WGS 84
Datum)}
\item{Advisory}{CyanoHAB advisory level for the incident as \href{https://mywaterquality.ca.gov/monitoring_council/cyanohab_network/docs/2016/appendix_a_2016_1.pdf}{defined by CCHAB}.
Is one of the following: Caution, Warning, or Danger.}
}
}
\source{
The HAB Incident Reports Map maintained by the State Water Resources
Control Board:
\url{https://mywaterquality.ca.gov/habs/where/freshwater_events.html}
}
\usage{
hab_incidents
}
\description{
A data set containing CyanoHAB incident reports in the upper San
Francisco Estuary (Delta) in 2021. Most of this data is from the State
Water Board's HAB Incident Report Map. This is combined with the maximum
advisory levels for Microcystins that occurred at each station in 2021 from
the cyanotoxin concentration data in \code{\link{hab_toxins}}. The \href{https://mywaterquality.ca.gov/monitoring_council/cyanohab_network/docs/2016/appendix_a_2016_1.pdf}{advisory levels}
were developed by the California Cyanobacteria and Harmful Algal Bloom
Network (CCHAB). Used in the analyses for the Spring-Summer version of the
2022 HABs/Weeds report for the Emergency Drought Barrier.
}
\keyword{datasets}
