% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hab_sat_ci_edb_reg}
\alias{hab_sat_ci_edb_reg}
\title{Daily average Cyano Index values for the EDB regions}
\format{
data frame with 1278 rows and 3 columns
\describe{
\item{Date}{Date of satellite imagery}
\item{Region}{Name of the EDB region. Either Sacramento, San Joaquin, or
Central Delta.}
\item{mean_pixel_val}{Daily average Cyano Index pixel value for the area in \code{Region}}
}
}
\source{
San Francisco Estuary Institute (SFEI). Harmful Algal Blooms Analysis
Tool. \url{https://fhab.sfei.org/}
}
\usage{
hab_sat_ci_edb_reg
}
\description{
A data set containing daily average Cyano Index pixel values for
the Emergency Drought Barrier (EDB) regions for the spring through late
fall (May-Dec) in 2020 and 2021. Zonal statistics were calculated from
satellite data downloaded from the Harmful Algal Blooms Analysis Tool (see
Source section below). Used in the analyses for the EDB.
}
\keyword{datasets}
