% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hab_sat_ci_fr_mil}
\alias{hab_sat_ci_fr_mil}
\title{Daily average Cyano Index values for Franks Tract and Mildred Island}
\format{
data frame with 852 rows and 3 columns
\describe{
\item{Date}{Date of satellite imagery}
\item{Name}{Name of the polygon area. Either Franks Tract or Mildred
Island.}
\item{mean_pixel_val}{Daily average Cyano Index pixel value for the area in \code{Name}}
}
}
\source{
San Francisco Estuary Institute (SFEI). Harmful Algal Blooms Analysis
Tool. \url{https://fhab.sfei.org/}
}
\usage{
hab_sat_ci_fr_mil
}
\description{
A data set containing daily average Cyano Index pixel values for
Franks Tract and Mildred Island for the spring through late fall (May-Dec)
in 2020 and 2021. Zonal statistics were calculated from satellite data
downloaded from the Harmful Algal Blooms Analysis Tool (see Source section
below). Used in the analyses for the Emergency Drought Barrier (EDB).
}
\keyword{datasets}
