% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hab_sat_fr_mil}
\alias{hab_sat_fr_mil}
\title{Summary of Cyano Index values derived from HAB satellite data for
Franks Tract and Mildred Island}
\format{
data frame with 329 rows and 9 columns
\describe{
\item{Date}{Date of satellite imagery}
\item{Name}{Name of the polygon region. Either Franks Tract or Mildred
Island.}
\item{AvgCI}{Average Cyano Index value of all valid pixels within the
region. The Cyano Index is a exponential and unitless metric.}
\item{NonDetect}{Count of the number of pixels below the detection limit
for the method which is less than or equal to a Cyano Index of 0.0000631}
\item{Low}{Count of the number of pixels in the Low Cyano Index category
defined as having a Cyano Index of greater than 0.00006310 but less than
0.0002}
\item{Moderate}{Count of the number of pixels in the Moderate Cyano Index
category defined as having a Cyano Index greater than or equal to 0.0002
but less than 0.001}
\item{High}{Count of the number of pixels in the High Cyano Index category
defined as having a Cyano Index greater than or equal to 0.001 but less
than 0.01}
\item{VeryHigh}{Count of the number of pixels in the Very High Cyano Index
category defined as having a Cyano Index greater than or equal to 0.01. The
maximum detectable level for the Cyano Index is 0.06327.}
\item{InvalidOrMissing}{Count of the number of pixels that were either
invalid (land, adjacency) or missing (clouds, no data, or some other
reason)}
}
}
\source{
San Francisco Estuary Institute (SFEI). Harmful Algal Blooms Analysis
Tool. \url{https://fhab.sfei.org/}
}
\usage{
hab_sat_fr_mil
}
\description{
A data set containing counts and averages of Cyano Index values
from HAB satellite data downloaded from the Harmful Algal Blooms Analysis
Tool (see Source section below). Summary statistics cover the spring
through late fall (May-Dec) in 2020 and 2021 for two open water regions in
the upper San Francisco Estuary: Franks Tract and Mildred Island. Used in
the analyses for the Emergency Drought Barrier (EDB).

The data set contains averages of Cyano Index for each region and date in
addition to pixel counts within 4 Cyano Index categories. The Cyano Index
categories (Low, Moderate, High, and Very High) were based on WHO
recreational guidance level thresholds. The data set also includes counts
of pixels that were below the detection limit for the imagery processing
method and counts of pixels that were either invalid or missing. Zeros were
substituted for the values below the detection limit and invalid or missing
pixels were excluded when calculating the average.

Counts and averages only include pixels that were completely within the
polygons for the two regions. In addition, the data set only includes
date-region combinations for when there were greater than 25\% valid pixels
within the region.
}
\keyword{datasets}
