% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hab_sat_ow_delta}
\alias{hab_sat_ow_delta}
\title{Counts of pixel values within Cyano Index categories for four open
water regions in the Delta}
\format{
data frame with 800 rows and 8 columns
\describe{
\item{Date}{Date of satellite imagery in yyyy-mm-dd}
\item{Region}{Name of the polygon region. Is one of the following: Franks
Tract, Mildred Island, Clifton Court Forebay, or Liberty Island.}
\item{NonDetect}{Count of the number of pixels below the detection limit
for the method which is less than or equal to a Cyano Index of 0.0000631}
\item{Low}{Count of the number of pixels in the Low Cyano Index category
defined as having a Cyano Index of greater than 0.00006310 but less than
0.0002}
\item{Moderate}{Count of the number of pixels in the Moderate Cyano Index
category defined as having a Cyano Index greater than or equal to 0.0002
but less than 0.001}
\item{High}{Count of the number of pixels in the High Cyano Index category
defined as having a Cyano Index greater than or equal to 0.001 but less
than 0.01}
\item{VeryHigh}{Count of the number of pixels in the Very High Cyano Index
category defined as having a Cyano Index greater than or equal to 0.01. The
maximum detectable level for the Cyano Index is 0.06327.}
\item{InvalidOrMissing}{Count of the number of pixels that were either
invalid (land, adjacency) or missing (clouds, no data, or some other
reason)}
}
}
\source{
San Francisco Estuary Institute (SFEI). Harmful Algal Blooms Analysis
Tool. \url{https://fhab.sfei.org/}
}
\usage{
hab_sat_ow_delta
}
\description{
A data set containing counts of pixel values derived from HAB
satellite data within 4 Cyano Index categories for four open water regions
in the upper San Francisco Estuary (Delta): Franks Tract, Mildred Island,
Clifton Court Forebay, and Liberty Island. Summary statistics cover the
summer through fall (June-Oct) in 2019-2021. The Cyano Index categories
(Low, Moderate, High, and Very High) were based on WHO recreational
guidance level thresholds. The data set also includes counts of pixels that
were below the detection limit for the imagery processing method and counts
of pixels that were either invalid or missing.

Counts only include pixels that were completely within the polygons for the
four regions. In addition, the data set only includes date-region
combinations for when there were greater than 25\% valid pixels within the
region. Zonal statistics were calculated from satellite data downloaded
from the Harmful Algal Blooms Analysis Tool (see Source section below).
Used in the analyses for the Emergency Drought Barrier (EDB).
}
\keyword{datasets}
