% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_vis_index_wq}
\alias{mc_vis_index_wq}
\title{\emph{Microcystis} visual index values and water quality measurements in
the upper San Francisco Estuary}
\format{
data frame with 7,339 rows and 11 columns
\describe{
\item{Source}{Name of the source dataset. Is one of the following: DWR_EMP
(DWR's Environmental Monitoring Program), DWR_NCRO (DWR's North Central
Region Office), FMWT (CDFW's Fall Midwater Trawl survey), or STN (CDFW's
Summer Townet survey)}
\item{Station}{Location where measurement was collected}
\item{Latitude}{Latitude of \code{Station} in Decimal Degrees (WGS 84 Datum)}
\item{Longitude}{Longitude of \code{Station} in Decimal Degrees (WGS 84 Datum)}
\item{Region}{Region designation of \code{Station} for the Emergency Drought
Barrier analysis. Is one of the following: Cache/Liberty, East Delta,
Franks, Lower Sac, Lower SJ, OMR, South Delta, or Upper Sac.}
\item{Year}{Calendar year of the measurement}
\item{Month}{Numeric calendar month of the measurement where 1 represents
January and 12 represents December}
\item{Date}{Calendar date of the measurement in yyyy-mm-dd}
\item{Microcystis}{\emph{Microcystis} visual index value on a qualitative scale
from 1 to 5, where 1 = absent, 2 = low, 3 = medium, 4 = high, and 5 = very
high}
\item{Secchi}{Secchi depth in centimeters}
\item{Temperature}{Water temperature in degrees Celsius}
}
}
\source{
\itemize{
\item \code{discretewq} data package (version 2.3.2):
\url{https://github.com/sbashevkin/discretewq/tree/v2.3.2}
\item California Department of Water Resources (DWR) Water Data Library (WDL):
\url{https://wdl.water.ca.gov/WaterDataLibrary/}
}
}
\usage{
mc_vis_index_wq
}
\description{
A data set containing \emph{Microcystis} visual index, secchi depth,
and water temperature data collected at various locations in the upper San
Francisco Estuary (Delta) from 2007-2021. Used in the analyses for the
Spring-Summer version of the 2022 HABs/Weeds report for the Emergency
Drought Barrier.

All \emph{Microcystis} visual index and water quality measurement data from
DWR_EMP (DWR's Environmental Monitoring Program), FMWT (CDFW's Fall
Midwater Trawl survey), and STN (CDFW's Summer Townet survey)  are from the
\code{discretewq} data package. All data collected by DWR_NCRO (DWR's North
Central Region Office) was acquired through a direct data request; however,
most of this data is available for download from the Water Data Library
(WDL).
}
\keyword{datasets}
