% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phyto_hab}
\alias{phyto_hab}
\title{Phytoplankton community data of potentially toxic Cyanobacteria in the
upper San Francisco Estuary}
\format{
data frame with 228 rows and 11 columns
\describe{
\item{Station}{Location where sample was collected}
\item{Region}{Region designation of \code{Station} for the Emergency Drought
Barrier analysis. Is one of the following: East Delta, Franks, Lower Sac,
Lower SJ, OMR, South Delta, or Upper Sac.}
\item{Year}{Calendar year of the sample}
\item{Date}{Calendar date of the sample in yyyy-mm-dd}
\item{Datetime}{Date and time (yyyy-mm-dd HH:MM:SS) of the sample in PST}
\item{Taxon}{Taxonomic designation of the organism as "Genus species"}
\item{Genus}{Genus designation of the organism. The genus \emph{Anabaena} is
grouped with \emph{Dolichospermum}.}
\item{Species}{Species designation of the organism}
\item{AlgalType}{General Algal Type classification of the organism}
\item{Count}{Number of organisms found in the sample}
\item{OrganismsPerMl}{Estimated number of organisms per milliliter of
sample}
}
}
\source{
Acquired from DWR's Environmental Monitoring Program through direct
data requests
}
\usage{
phyto_hab
}
\description{
A data set containing phytoplankton community data of
potentially toxic Cyanobacteria collected by DWR's Environmental Monitoring
Program at various locations in the upper San Francisco Estuary (Delta)
from 2014-2021. Used in the analyses for the Spring-Summer version of the
2022 HABs/Weeds report for the Emergency Drought Barrier.
}
\keyword{datasets}
