% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sat_ci_count_fr_mil}
\alias{sat_ci_count_fr_mil}
\title{Counts of pixel values within Cyano Index categories for Franks Tract
and Mildred Island}
\format{
data frame with 329 rows and 3 columns
\describe{
\item{Date}{Date of satellite imagery}
\item{Name}{Name of the polygon region. Either Franks Tract or Mildred
Island.}
\item{Non_detect}{Count of the number of pixels below the detection limit
for the method which is less than or equal to 6,310 cells/mL}
\item{Low}{Count of the number of pixels in the Low Cyano Index category
defined as having greater than 6,310 cells/mL but less than 19,999
cells/mL}
\item{Moderate}{Count of the number of pixels in the Moderate Cyano Index
category defined as having between 19,999 and 99,999 cells/mL}
\item{High}{Count of the number of pixels in the High Cyano Index category
defined as having between 99,999 and 999,999 cells/mL}
\item{Very_high}{Count of the number of pixels in the Very High Cyano Index
category defined as having greater than 999,999 cells/mL}
\item{Invalid_or_missing}{Count of the number of pixels that were either
invalid (land, adjacency) or missing (clouds, no data, or some other
reason)}
}
}
\source{
San Francisco Estuary Institute (SFEI). Harmful Algal Blooms Analysis
Tool. \url{https://fhab.sfei.org/}
}
\usage{
sat_ci_count_fr_mil
}
\description{
A data set containing counts of pixel values within 5 Cyano
Index categories for Franks Tract and Mildred Island for the spring through
late fall (May-Dec) in 2020 and 2021. The Cyano Index categories were based
on WHO recreational guidance level thresholds. Counts only include pixels
that were completely within the polygons for the two regions. The data set
only includes date-region combinations where there were greater than 25\%
valid pixels within the region. Zonal statistics were calculated from
satellite data downloaded from the Harmful Algal Blooms Analysis Tool (see
Source section below). Used in the analyses for the Emergency Drought
Barrier (EDB).
}
\keyword{datasets}
