% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wtr_temp_daily_dd}
\alias{wtr_temp_daily_dd}
\title{Daily average water temperature and degree days for the Central Delta}
\format{
data frame with 2,557 rows and 4 columns
\describe{
\item{Year}{Calendar year of the measurement}
\item{Date}{Calendar date of the measurement in yyyy-mm-dd}
\item{WaterTemp_Avg}{Daily average water temperature in degrees Celsius of
six continuous water quality stations located in the Central and Southern
Delta}
\item{WaterTemp_DD}{Daily accumulated degree days above 19C for the
averaged water temperature data at six stations located in the Central and
Southern Delta for each \code{Year}. Calculated with the \code{\link[pollen:gdd]{pollen::gdd()}}
function.}
}
}
\source{
\itemize{
\item California Department of Water Resources (DWR) Water Data Library (WDL):
\url{https://wdl.water.ca.gov/WaterDataLibrary/}
\item USGS National Water Information System (NWIS):
\url{https://nwis.waterdata.usgs.gov/nwis}
}
}
\usage{
wtr_temp_daily_dd
}
\description{
A data set containing daily average water temperatures and
degree days above 19 degrees Celsius for the upper San Francisco Estuary
(Delta) from 2015-2021. Values are based on averages of six continuous
water quality stations located in the Central and Southern Delta: False
River near Oakley (FAL), Middle River near Holt (HLT), Holland Cut near
Bethel Island (HOL), Middle River at Middle River (MDM), Old River at
Quimbly (ORQ), and Old River at Franks Tract near Terminous (OSJ). Used in
the analyses for the Spring-Summer version of the 2022 HABs/Weeds report
for the Emergency Drought Barrier.

All water temperature data collected by NCRO (DWR's North Central Region
Office; FAL, HLT, HOL, ORQ, and OSJ stations) was downloaded from the Water
Data Library (WDL). Data collected by USGS (MDM station) was downloaded
from the National Water Information System (NWIS).
}
\keyword{datasets}
