% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cont_chla_daily}
\alias{cont_chla_daily}
\title{Daily values of continuous chlorophyll fluorescence data for stations
within the EDB regions}
\format{
data frame with 7,612 rows and 7 columns
\describe{
\item{Source}{Name of the source dataset. Is one of the following: DWR_CEMP
(DWR's Continuous Environmental Monitoring Program), DWR_NCRO (DWR's North
Central Region Office), or USGS}
\item{Station}{Location where measurement was collected}
\item{Region}{Region designation of \code{Station} for the Emergency Drought
Barrier analysis. Is one of the following: Central Delta, Sacramento, or
San Joaquin.}
\item{Year}{Calendar year of the measurement}
\item{Date}{Calendar date of the measurement in yyyy-mm-dd}
\item{AvgChla}{Daily average chlorophyll fluorescence value in micrograms
per liter}
\item{MedianChla}{Daily median chlorophyll fluorescence value in micrograms
per liter}
}
}
\source{
\itemize{
\item California Department of Water Resources (DWR) Water Data Library (WDL):
\url{https://wdl.water.ca.gov/WaterDataLibrary/}
\item USGS National Water Information System (NWIS):
\url{https://nwis.waterdata.usgs.gov/nwis}
}
}
\usage{
cont_chla_daily
}
\description{
A data set containing daily average and median values of
continuous chlorophyll fluorescence measured in the water in-situ at the
stations within the designated Emergency Drought Barrier (EDB) regions.
Data was collected by DWR and USGS from 2020-2021. Used in the analyses for
the EDB.

All chlorophyll data collected by DWR_NCRO (DWR's North Central Region
Office) was downloaded from the Water Data Library (WDL). Data collected by
DWR_CEMP (DWR's Continuous Environmental Monitoring Program) was acquired
through direct data requests. And data from USGS were downloaded directly
from the National Water Information System (NWIS) using the \code{dataRetrieval}
R package. Some of the data from USGS is considered provisional.
}
\keyword{datasets}
